Add missing fields to RT spec (#145)
RT spec is missing
- some properties related to multichannel diarization: `channel_ids` in
the transcript output, `channel_diarization_labels` in the transcription
config, `channel` and `channel_and_speaker` enum value for diarization
in transcription config.
- fields related to entities, returned when `enable_entites` is `true`
in the transcription config

Also, in the transcript output, `channel` is not defined at each
`result` level. It is defined at "root level" for each
`AddTranscript`/etc message.

This MR syncs the internal spec file with the github version. Other
properties not included in this PR are those related to "force end of
utterance" feature which have their own PR
https://github.com/speechmatics/docs/pull/135

Co-authored-by: Claire Schaefer <claires@speechmatics.com>